+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to WindowRecorder will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.7] - 2026-07-15
+
+### Added
+- Popup dialog when recording completes with "Open Video", "Show in Finder", "Dismiss"
+- Dialog activates app to foreground on Tahoe
+
 ## [1.0.6] - 2026-07-15
 
 ### Fixed